Ensure render-page does not swallow exception #27
If we're processing an exception, render-page should produce a 500 error
with the body being a string representation of the exception instead of
a 404 not found, so we now pass a flag into render-page to determine the
behavior when the page is not found.

This feels a bit hacky so revisiting and refactoring this at some point
would be worthwhile.