Jetty server should not join by default
http-kit spins up on a separate thread and we deref a promise after the
stop function runs in order to wait for it. Jetty can do this but its
default is :join? true so it waits internally. Now we tell it not to
wait internally so the same machinery works for both servers.